requirements.txt
woocommerce
requests
beautifulsoup4
pandas
openpyxl
python-dotenv